should be a long, not an int
git-svn-id: https://svn.eu.apache.org/repos/asf/incubator/abdera/java/trunk@629227 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/abdera/factory/StreamBuilder.java
@@ -354,7 +354,7 @@ public class StreamBuilder
   }
 
   public StreamBuilder startLink(String iri, String rel, String type,
-      String title, String hreflang, int length) {
+      String title, String hreflang, long length) {
     return (StreamBuilder)super.startLink(iri, rel, type, title, hreflang, length);
   }
 
@@ -721,7 +721,7 @@ public class StreamBuilder
   }
 
   public StreamBuilder writeLink(String iri, String rel, String type,
-      String title, String hreflang, int length) {
+      String title, String hreflang, long length) {
     return (StreamBuilder)super.writeLink(iri, rel, type, title, hreflang, length);
   }
